test_app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "fastapi-cors>=0.0.6",
    "psycopg2-binary>=2.9.11",
    "uvicorn[standard]>=0.42.0",
]
